backend/requirements.txt
alembic==1.4.2
backcall==0.2.0
click==7.1.2
decorator==4.4.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
ipython==7.18.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
parso==0.7.1
passlib==1.7.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
psycopg2-binary==2.8.6
ptyprocess==0.6.0
Pygments==2.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.19
traitlets==5.0.4
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
